Stop sys.path wrangling
