+++ scripts/preprocessing/03-make_epochs.py
@@ -17,9 +17,8 @@ import mne
 from mne.parallel import parallel_func
 from mne_bids import BIDSPath
 
-from common import make_epochs
 import config
-from config import gen_log_message, on_error, failsafe_run
+from config import make_epochs, gen_log_message, on_error, failsafe_run
 
 logger = logging.getLogger('mne-bids-pipeline')
